+++ lib/remote_config_gen.dart
@@ -3,5 +3,6 @@ library remote_config_gen;
 
 export 'src/remote_config_generator.dart';
 export 'src/models/generation_config.dart';
+export 'src/models/converter_config.dart';
 export 'src/models/remote_config_data.dart';
 export 'src/exceptions/remote_config_exception.dart';